requirements.txt
# Django and REST framework
Django>=5.1.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.27.1
django-debug-toolbar>=4.3.0
django-extensions>=3.2.3

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Testing
pytest>=8.3.5
pytest-django>=4.11.1
factory-boy>=3.3.0
Faker>=37.1.0
requests>=2.32.3

# Development tools
black>=24.2.0  # Code formatting
flake8>=7.0.0  # Code linting
isort>=5.13.2  # Import sorting

# API documentation
drf-spectacular>=0.27.1  # OpenAPI schema generation

# Utilities
python-dotenv>=1.0.1  # Environment variable management
Pillow>=10.2.0  # Image processing 